rename loader.lua to android.lua
you can use require("android") - but note that this will NOT actually
load the android.lua file, since this has been done already by the
JNI/C launcher code. Rather, the require will just return what we
have already stored in the packages.loaded table in the Lua environment.

This way, we can use the "android" module in our own code as well as the
initial Lua loader (which is also in android.lua).