[fix] Normalize Android build
There are different numbers all over the place, but I *think* the code points toward Android 4 (platform-14) as an intended target.

Also, by using `NDKABI=${NDKABI:-14}` we allow for environment variable overrides. It only sets the default of 14 if $NDKABI isn't set. This is in the same spirit as https://github.com/koreader/koreader/pull/3064