Add an eink-test activity (#98)
This isn't built as part of android-luajit-launcher. It's a different apk which comes in handy to test
currently implemented EPD controllers. It provides info about device and lets the user to quickly
test if the device can be added to KOReader.

It is safe to run this activity on devices without e-ink displays, `adb logcat eink-test:V *:S` reports:

```adb logcat
01-30 16:20:32.932 20165 20165 I eink-test: fake update
01-30 16:20:33.885 20165 20165 I eink-test: rockchip forced update
01-30 16:20:33.886 20165 20165 E eink-test: java.lang.NullPointerException: enumType == null
01-30 16:22:41.853 20590 20590 I eink-test: fake update
01-30 16:22:42.643 20590 20590 I eink-test: rockchip normal update
01-30 16:22:42.646 20590 20590 E eink-test: java.lang.NullPointerException: enumType == null
```