Attempt to modernize the mcode alloc hackery...
AFAICT, the mmap trick is *way too late* to be any good, so, get rid of
it, and mention that a proper approach would require moving it to the C
frontend, which in turn would require a shared build of LuaJIT, instead
of embedding it statically.

Also, use modern syntax for jit.opt, and update to the current defaults,
to actually make sure it won't try to grow it later.