Removed ONYX Book GO7 and GO COLOR 7 From Onyx/Qualcomm EDP
This is in response to @jadehawk fork of the same file. I've removed Go Color 7 as well, for the exact same reason, namely:

"Due to intermittent black screen when device awakes from sleep. I was able to duplicate the issue consistently when the EDP driver was assigned to Onyx/Qualcomm even if it did work on the developer Test Page."